+++ crates/rpc/src/server/mod.rs
@@ -80,7 +80,11 @@ impl Rpc {
             .layer(CatchPanicLayer::custom(catch_panic_layer_fn))
             .layer(TraceLayer::new_for_grpc().make_span_with(grpc_trace_fn))
             .layer(HealthCheckLayer)
-            .layer(AcceptHeaderLayer::new(&rpc_version, genesis.commitment()))
+            .layer(
+                AcceptHeaderLayer::new(&rpc_version, genesis.commitment())
+                    .with_genesis_enforced_method("SubmitProvenTransaction")
+                    .with_genesis_enforced_method("SubmitProvenBatch"),
+            )
             .layer(cors_for_grpc_web_layer())
             // Enables gRPC-web support.
             .layer(GrpcWebLayer::new())